Updated to Visual Studio 2015 Preview. Dropped XP support.  No longer using 2010 DirectX SDK.
git-svn-id: http://zigserv/svn/Projects/trunk/wcdx@384 41632dc1-7a16-0410-8441-a1f66f767317

+++ src/wcdx.cpp
@@ -144,7 +144,7 @@ HRESULT STDMETHODCALLTYPE Wcdx::UpdatePalette(UINT index, const WcdxColor* entry
 
 HRESULT STDMETHODCALLTYPE Wcdx::UpdateFrame(INT x, INT y, UINT width, UINT height, UINT pitch, const byte* bits)
 {
-	RECT rect = { x, y, x + width, y + height };
+	RECT rect = { x, y, LONG(x + width), LONG(y + height) };
 	RECT clipped =
 	{
 		max(rect.left, LONG(0)),
